requirements_test.txt
pytest-homeassistant-custom-component==0.13.175
webauthn==2.2.0
josepy<2.0
